Move code to src main folder
