+++ src/components/navigation/TabBarIcon.tsx
@@ -1,5 +1,4 @@
-// You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
-
+import React from "react";
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { type IconProps } from '@expo/vector-icons/build/createIconSet';
 import { type ComponentProps } from 'react';
