+++ src/hooks/useThemeColor.ts
@@ -5,7 +5,7 @@
 
 import { useColorScheme } from 'react-native';
 
-import { Colors } from '@/constants/Colors';
+import { Colors } from '@/src/constants/Colors';
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
