+++ src/scripts/reset-project.js
